Got rid of single class folders
