--- src/main/java/frc/robot/subsystems/Endgame/Endgame.java
@@ -1,5 +1,0 @@
-package frc.robot.subsystems.endgame;
-
-public class Endgame {
-    
-}
