+++ src/main/java/frc/robot/subsystems/Vision.java
@@ -1,0 +1,5 @@
+package frc.robot.subsystems;
+
+public class Vision {
+    
+}
