--- src/main/java/frc/robot/subsystems/Vision/Vision.java
@@ -1,5 +1,0 @@
-package frc.robot.subsystems.vision;
-
-public class Vision {
-    
-}
