--- src/main/java/frc/robot/subsystems/gpmSubsystem/Intake/Intake.java
@@ -1,5 +1,0 @@
-package frc.robot.subsystems.gpmSubsystem.intake;
-
-public class Intake {
-    
-}
