--- src/main/java/frc/robot/subsystems/gpmSubsystem/Outtake/Outtake.java
@@ -1,5 +1,0 @@
-package frc.robot.subsystems.gpmSubsystem.outtake;
-
-public class Outtake {
-    
-}
