--- src/main/java/frc/robot/subsystems/gpmSubsystem/StorageIndex/StorageIndex.java
@@ -1,5 +1,0 @@
-package frc.robot.subsystems.gpmSubsystem.storageIndex;
-
-public class StorageIndex {
-    
-}
