+++ src/main/java/frc/robot/subsystems/gpm_subsystem/Outtake.java
@@ -1,0 +1,5 @@
+package frc.robot.subsystems.gpm_subsystem;
+
+public class Outtake {
+    
+}
